feat(install): add GitHub archive and CLI helpers with improved async handling
- Add GitHubArchiveHelper utility for handling GitHub archive downloads
- Add GitHubCLIHelper utility for GitHub CLI operations
- Register new helpers in dependency injection container
- Convert window dimension properties from double to int in AppConfig for consistency
- Add ConfigureAwait(false) to all Task.Run calls for better async performance
- Update YunzaiInstallService to use new archive and CLI helpers
- Increase YunzaiInstallService installation steps from 6 to 11 for better progress tracking
- Improve async/await patterns across all install services (AstrBot, DDBot, Koishi, Yunzai, Zhenxun)
- Update imports to use new Utils namespace for GitHub helpers

+++ Models/AppConfig.cs
@@ -55,16 +55,16 @@ public class AppConfig
     public string ThemeMode { get; set; } = "dark";
 
     [JsonPropertyName("window_width")]
-    public double WindowWidth { get; set; } = 1200.0;
+    public int WindowWidth { get; set; } = 1200;
 
     [JsonPropertyName("window_height")]
-    public double WindowHeight { get; set; } = 800.0;
+    public int WindowHeight { get; set; } = 800;
 
     [JsonPropertyName("window_left")]
-    public double? WindowLeft { get; set; } = null;
+    public int? WindowLeft { get; set; } = null;
 
     [JsonPropertyName("window_top")]
-    public double? WindowTop { get; set; } = null;
+    public int? WindowTop { get; set; } = null;
 
     // 兼容性属性 - 只读取不写入
     [JsonPropertyName("auto_login")]
